+++ src/api/getGhostPosts.ts
@@ -1,24 +1,13 @@
-import { THREADS_API_BASE_URL } from "../constants.ts";
+import { THREADS_API_BASE_URL, USER_THREADS_FIELDS } from "../constants.ts";
 import type { PaginationOptions, ThreadsListResponse } from "../types.ts";
 import { getAPI } from "../utils/getAPI.ts";
 
 const DEFAULT_FIELDS = [
-	"id",
-	"media_product_type",
-	"media_type",
-	"media_url",
-	"permalink",
-	"username",
-	"text",
-	"timestamp",
-	"shortcode",
-	"thumbnail_url",
-	"children",
-	"is_quote_post",
-	"alt_text",
-	"link_attachment_url",
-	"has_replies",
-	"reply_audience",
+	...USER_THREADS_FIELDS,
+	"poll_attachment",
+	"topic_tag",
+	"ghost_post_status",
+	"ghost_post_expiration_timestamp",
 ];
 
 /**